python/requirements.txt
asyncpg
gunicorn
uvicorn-worker